fix: address 6 critical/high findings from security review round 2
core.py: add float handler in env var coercion (prevents OOM on string * 10^18),
  add ISSUER_URLS + ISSUER_DISCOVERY to env_map (config was lying)
bot.py: cast stake_wei to int (float * 10^18 rejected by web3.py)
swarm-api.ts: nonce++ after await, re-fetch on failure (prevents desync)
deploy-swarm.sh: stop before build, SSH timeouts, pnl-data dir ownership

+++ vision-bot/bot.py
@@ -134,7 +134,7 @@ def run_cycle(cfg, executor, tracker, strategy, risk, issuer_urls_fn, feed):
         # Dual-balance flow: deposit USDC into Vision balance, then joinBatch pulls from it
         executor.approve_usdc(deposit_wei)
         executor.deposit_balance(deposit_wei)
-        stake_wei = cfg["stake"] * 10**DECIMALS
+        stake_wei = int(cfg["stake"] * 10**DECIMALS)
         executor.join_batch(batch_id, config_hash, deposit_wei, stake_wei, bm_hash)
 
         time.sleep(2)  # initial wait for block confirmation
